Crate advanca_core

Structs

Enclave
The public information about an Enclave

Module
The module declaration.

PublicKeys

Task
Task information

TaskSpec
The task specification. //TODO: more details

User
The public information about a user

Worker
The public information about a worker and its enclave

Enums

Call
Dispatchable calls.

Error

Privacy
The possible privacy levels

RawEvent
Events for this module.

TaskStatus
The status of a task managed on chain

Traits

Trait
The module's configuration trait.

Functions

sr25519_verify_msg

sr25519_verify_signature

Type Definitions

AccountId
Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex
The type for looking up accounts. We don't expect more than 4 billion of them, but you never know...

Balance
Balance of an account.

BalanceOf

BlockNumber
An index to a block.

Ciphertext
The encrypted bytes

DigestItem
Digest item type.

Duration
Duration of the task. 0 means unlimited.

Event
RawEvent specialized for the configuration Trait

Hash
A hash of some data used by the chain.

Index
Index of a transaction in the chain.

Index

Signature
Alias to 512-bit hash when used in the context of a transaction signature on the chain.

TaskId